add duration and transfer size features

+++ src/tests/features/features.h
@@ -8,6 +8,8 @@
 #ifndef SRC_TESTS_FEATURES_FEATURES_H_
 #define SRC_TESTS_FEATURES_FEATURES_H_
 
-typedef double feature_type_t;
+#include <stdint.h>
+
+typedef uint64_t feature_type_t;
 
 #endif /* SRC_TESTS_FEATURES_FEATURES_H_ */
